Meal'a yemek eklenirken 100'e tam bölünmeyen gram girildiğinde yanlış hesaplama sorunu giderildi. Create Food formundan Form1'e dönlülürken 1den fazla Form1 oluşma sorunu giderildi

+++ CalCalculatorBLL/MealServices.cs
@@ -59,7 +59,7 @@ namespace CalCalculatorBLL
                new
                {
                    FoodName = x.Food.FoodName,
-                   Calories = x.Food.FoodCal * (x.Grams / 100),
+                   Calories = x.Food.FoodCal * (double)(x.Grams / 100.00),
                    Grams = x.Grams,
                    FoodID = x.FoodID
